# Conflicts:
#	Source/Lib/Common/Codec/EbDefinitions.h
#	Source/Lib/Common/Codec/EbResourceCoordinationProcess.c

+++ Source/App/EncApp/EbAppMain.c
@@ -258,24 +258,24 @@ int32_t main(int32_t argc, char* argv[])
                     if (configs[instanceCount]->stop_encoder == EB_FALSE) {
                         // Interlaced Video
                         if (configs[instanceCount]->interlaced_video || configs[instanceCount]->separate_fields) {
-                            printf("\nChannel %u\nAverage Speed:\t\t%.0f fields per sec\nTotal Encoding Time:\t\t%.0f ms\nTotal Execution Time:\t\t%.2f ms\nAverage Latency:\t%.0f ms\nMax Latency:\t\t%u ms\nCycle Count:\t\t%llu cycles\n",
+                            printf("\nChannel %u\nAverage Speed:\t\t%.0f fields per sec\nTotal Encoding Time:\t\t%.0f ms\nTotal Execution Time:\t\t%.2f ms\nAverage Latency:\t%.0f ms\nMax Latency:\t\t%u ms\nCycle Count:\t\t%lu cycles\n",
                                 (uint32_t)(instanceCount + 1),
                                 configs[instanceCount]->performance_context.average_speed,
                                 configs[instanceCount]->performance_context.total_encode_time * 1000,
                                 configs[instanceCount]->performance_context.total_execution_time * 1000,
                                 configs[instanceCount]->performance_context.average_latency,
                                 (uint32_t)(configs[instanceCount]->performance_context.max_latency),
-                                (configs[instanceCount]->performance_context.end_cycle_count - configs[instanceCount]->performance_context.start_cycle_count));
+                                (unsigned long)(configs[instanceCount]->performance_context.end_cycle_count - configs[instanceCount]->performance_context.start_cycle_count));
                         }
                         else {
-                            printf("\nChannel %u\nAverage Speed:\t\t%.3f fps\nTotal Encoding Time:\t%.0f ms\nTotal Execution Time:\t%.0f ms\nAverage Latency:\t%.0f ms\nMax Latency:\t\t%u ms\nCycle Count:\t\t%llu cycles\n",
+                            printf("\nChannel %u\nAverage Speed:\t\t%.3f fps\nTotal Encoding Time:\t%.0f ms\nTotal Execution Time:\t%.0f ms\nAverage Latency:\t%.0f ms\nMax Latency:\t\t%u ms\nCycle Count:\t\t%lu cycles\n",
                                 (uint32_t)(instanceCount + 1),
                                 configs[instanceCount]->performance_context.average_speed,
                                 configs[instanceCount]->performance_context.total_encode_time * 1000,
                                 configs[instanceCount]->performance_context.total_execution_time * 1000,
                                 configs[instanceCount]->performance_context.average_latency,
                                 (uint32_t)(configs[instanceCount]->performance_context.max_latency),
-                                (configs[instanceCount]->performance_context.end_cycle_count - configs[instanceCount]->performance_context.start_cycle_count));
+                                (unsigned long)(configs[instanceCount]->performance_context.end_cycle_count - configs[instanceCount]->performance_context.start_cycle_count));
                         }
                     }
                     else
